% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_group.R
\name{bed_group}
\alias{bed_group}
\title{Groupwise application of functions over bed file}
\usage{
bed_group(data, vars, funs, group)
}
\arguments{
\item{data}{Data.frame}

\item{vars}{list of variable names. Must be supplied in the form
list(a, b, d).}

\item{funs}{List of functions to be applied to the region. See details for
more.}

\item{group}{Variable of denoting group}
}
\value{
The original data set with a column added for each fuction in funs.
}
\description{
Groupwise application of functions over bed file
}
\details{
The region that the funs will be applied to is a matrix with a column for
each sample and a row for each row.
}
\examples{
library(dplyr)
bed_sample <- mutate(bed_sample, ggg = ceiling(row_number() / 10))
bed_group(bed_sample,
          vars = c("beta1", "beta2", "beta3", "beta4", "beta5"),
          funs = list(length, mean),
          group = ggg)
}
