% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_slide.R
\name{bed_slide}
\alias{bed_slide}
\title{Sliding application of functions over bed file}
\usage{
bed_slide(data, vars, funs, .i, size)
}
\arguments{
\item{data}{Data.frame}

\item{vars}{list of variable names. Must be supplied in the form
list(a, b, d).}

\item{funs}{List of functions to be applied to the region. See details for
more.}

\item{.i}{Name of the variable denoting the location.}

\item{size}{Size of the sliding window. The window will be centered around
each site.}
}
\value{
The original data set with a column added for each fuction in funs.
}
\description{
Sliding application of functions over bed file
}
\details{
The region that the funs will be applied to is a matrix with a column for
each sample and a row for each row.
}
\examples{
bed_slide(bed_sample,
          vars = list(beta1, beta2, beta3, beta4, beta5),
          funs = list(length, mean),
          .i = start, size = 100)
}
