% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metric_mhic}
\alias{metric_mhic}
\title{Calculate MHIC metric for matrix}
\usage{
metric_mhic(x, lower = 0.2, upper = 0.8)
}
\arguments{
\item{x}{Numeric matrix}

\item{lower}{Numeric determining the lower bound.}

\item{upper}{Numeric determining the upper bound.}
}
\description{
Calculate MHIC metric for matrix
}
\examples{
mat <- matrix(seq(0.1, 0.9, by = 0.1), ncol = 3)
metric_mhic(mat)
}
