% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metric_pwd}
\alias{metric_pwd}
\title{Calculate Pairwise Distance metric for matrix}
\usage{
metric_pwd(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
Calculate Pairwise Distance metric for matrix
}
\examples{
mat <- matrix(1:9, ncol = 3)
metric_pwd(mat)
}
