% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_sample_var.R
\name{metric_sample_var}
\alias{metric_sample_var}
\title{Calculate sample variance metric for matrix}
\usage{
metric_sample_var(x)
}
\arguments{
\item{x}{Numeric matrix}
}
\value{
A numeric
}
\description{
Calculate sample variance metric for matrix
}
\examples{
mat <- matrix(seq(0.1, 0.9, by = 0.1), ncol = 3)
metric_sample_var(mat)
}
