% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bed.R
\name{read_bed}
\alias{read_bed}
\title{Read a .bed file}
\usage{
read_bed(file)
}
\arguments{
\item{file}{the name of the file which the data are to be read from.}
}
\value{
A [tibble][tibble::tibble-package] with 5 columns; "chr", "start",
"end", "beta" and "coverage".
}
\description{
Read a .bed file
}
