% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedslider-package.R
\docType{package}
\name{bedslider-package}
\alias{bedslider}
\alias{bedslider-package}
\title{bedslider: Calculate Metrics .bed Files}
\description{
Calculate metrics in a sliding fashion over .bed files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EmilHvitfeldt/bedslider}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/bedslider/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

}
\keyword{internal}
