% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metric_sample_var}
\alias{metric_sample_var}
\title{Calculate sample variance metric for matrix}
\usage{
metric_sample_var(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
Calculate sample variance metric for matrix
}
\examples{
mat <- matrix(1:9, ncol = 3)
metric_sample_var(mat)
}
